+++ website/src/AppRouter.tsx
@@ -5,8 +5,7 @@ import Home from "./Pages/home";
 import Login from "./Pages/login";
 import { EditController } from "./Pages/editController";
 import { Table } from "./Pages/table";
-import Profile from "./Components/Profile/profile";
-import Helmet from "./Components/Header/helmet";
+import Profile from "./Pages/profile";
 import PageHeader from "./Components/Header/pageHeader";
 import Footer from "./Components/Footer/footer";
 
@@ -14,10 +13,6 @@ const AppRouter = () => {
   return (
     <Router>
       <div className="flex flex-col min-h-screen bg-eastlake-grey font-text">
-        <Helmet
-          title="Login"
-          description="Login page where users can login by receiving an OTP link in their email"
-        />
         <PageHeader />
         <Switch>
           <Route path="/login">
