+++ website/src/Pages/login.tsx
@@ -1,9 +1,16 @@
 import React from "react";
 import LoginForm from "../Components/Login/loginForm";
+import Helmet from "../Components/Header/helmet";
 
 const Login = () => {
   return (
-    <LoginForm />
+    <>
+      <Helmet
+        title="Login"
+        description="Login page where users can login by receiving an OTP link in their email"
+      />
+      <LoginForm />
+    </>
   );
 };
