+++ website/src/Pages/table.tsx
@@ -1,18 +1,24 @@
 import React from "react";
 
 import { useSyncUserData } from "../Firebase/firestore/useUserData";
-
+import Helmet from "../Components/Header/helmet";
 import TableController from "../Components/Table/tableController";
 
 const Table: React.FC<{}> = () => {
   useSyncUserData();
   return (
-    <div className="mb-auto">
-      {
-        // mb - auto forces the footer to be at the bottom
-      }
-      <TableController />
-    </div>
+    <>
+      <Helmet
+        title="Table"
+        description="Page showing your current submitted hours."
+      />
+      <div className="mb-auto">
+        {
+          // mb - auto forces the footer to be at the bottom
+        }
+        <TableController />
+      </div>
+    </>
   );
 };
